Added files via upload
This is an spreadsheet that was output from the scraping script. Only the first 4 columns were filled out by the script as well as the last columns with no headers (NA). THis was because everything in the middle was meant to be filled out by a human; who's job was to inspect all the metadata associated with a project title and knb_id.